Fix xhp_preprocess_code in centos
zend_parse_parameters gives you an int, not a size_t. f u Zend.

Reviewed By: bill

Test Plan: build

+++ ext.cpp
@@ -285,7 +285,7 @@ ZEND_FUNCTION(__xhp_idx) {
 ZEND_FUNCTION(xhp_preprocess_code) {
   // Parse zend params
   char *code;
-  size_t code_len;
+  int code_len;
   if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "s", &code, &code_len) == FAILURE) {
     RETURN_NULL();
   }
